[OMPD] Make odb initialization at first command
ODB now initializes the openmp process with ompd when the first ompd
commadn is issued, not when the debugger starts.
This is necessary because we now check for state tracking in the
process initialization function, and OpenMP runtime may not be present
when the debugger has just loaded the program.